Adds averaging to ADC reads
This patch adds support for averaging the ADC reads
to improve the stability. This approach is preferable
to the current running-average solution because it is
completely internal to the driver and provides predictable
performance with low CPU overhead. The current patch only
support averaging for non-muxed ADC channels because I
have no way to test them.

Signed-off-by: Greg Burns <burns.greg@gmail.com>