Change config map to switch
Used to be array with indexing and if to
catch out of bounds error

This way we can just add a case

Should be noted, the review says the compiler will warn if we add
a new mode and don't update the switch.
I think this is untrue because of the default error condition.
Switchs only warn about missing cases if there's no default.