updated qspi to cpp standard
The compiled size on the example I used increased by 536 bytes, from 44,028 to 44,564.

Like i2c, this module takes a config struct for configuration. It may not align with other classes in `daisy_seed.h`.

I didn't test the CMake build, but it should work with my single change to the cmakelists file.