Fix builds on native Windows
Previously, builds would fail with a message stating that there's no such command "true".
CMAKE_C[XX]_ARCHIVE_FINISH is not a boolean option, but instead should be a string for how to run ranlib. This simply sets that to the default, but it could instead be removed in the future now that GCC binutils has lto automatically loading.