Replacing `true` with `""` in `*ARCHIVE_FINISH` directives
When setting up a new project, CMake will check to make sure the
compiler is working with the given toolchain. With the libDaisy
toolchain file, this test produces a command line invocation of qcs as
follows:

```
qcs libcmTC_a65c7.a  CMakeFiles/cmTC_a65c7.dir/testCCompiler.c.obj && true && cd .
```

When invoked in bash, this is a no-op and not an issue. However, on
Windows, this fails with the following message:

```
'true' is not recognized as an internal or external command
```

This change removes the `true` invocation with an empty string, and the
test will pass.