Improves reliability of multiple consecutive writes
before, the first write would be exact, but subsequent writes would introduce artifacts or inaccuracies to the wav file as compared to recording into a buffer. Powercycling would solve this but not ideal. Now, after 5 writes, there is usually not artefacts or differences between wav file and equivalent buffer. Although occasionally this isn't true so this isn't a complete fix and the issue should be looked into by someone who knows more about this class and fatfs than I.